requirements.txt

Django==1.11.1
gunicorn==19.6.0
psycopg2
pillow